Modified using split

+++ app/main.py
@@ -3,19 +3,19 @@ import os
 
 def move_file(command: str) -> None:
 
-    command_name, src_file, dest_file = command.split(" ")
+    command_list = command.split(" ")
 
-    if command_name == "mv" and os.path.exists(src_file):
+    if len(command_list) == 3:
+        command_name, src_file, dest_file = command_list
 
-        dest_list = dest_file.split("/")
-        path_dir = ""
+        if command_name == "mv" and os.path.exists(src_file):
 
-        for new_dir in dest_list[:-1]:
-            path_dir = os.path.join(path_dir, new_dir)
-            os.makedirs(path_dir, exist_ok=True)
+            path_dir, new_name = os.path.split(dest_file)
+            if path_dir:
+                os.makedirs(path_dir, exist_ok=True)
 
-        with (open(src_file, "r") as file_in,
-              open(dest_file, "w") as file_out):
-            file_out.write(file_in.read())
+            with (open(src_file, "r") as file_in,
+                  open(dest_file, "w") as file_out):
+                file_out.write(file_in.read())
 
-        os.remove(src_file)
+            os.remove(src_file)
